+++ dev/src/Business/ApplicationFacade/iControllers/iDriverRoleController.java
@@ -13,5 +13,5 @@ public interface iDriverRoleController {
 
     boolean checkAvailableStoreKeeperAndShifts(int BID,LocalDate date, LocalTime leavingTime);
 
-    void addDriverToShiftAndStoreKeeper(int BID,long driverID, LocalDate date, LocalTime leavingTime);
+    void addDriverToShiftAndStoreKeeper(int BID,int driverID, LocalDate date, LocalTime leavingTime);
 }
